изменил size_t на i

+++ src/task2.cpp
@@ -1,7 +1,7 @@
 #include <cmath>
 bool checkPrime(unsigned int value)
 {
-	for (size_t i = 2; i <= sqrt(value); i++)
+	for (int i = 2; i <= sqrt(value); i++)
 	{
 		if (value % i == 0)
 			return false;
